v1.6

+++ Scripts/FamilyRisk_single.sh
@@ -46,7 +46,9 @@ For filtering variants:
     --acmg-classification           User defined revel threshold, 
                                     default is Pathogenic, Likely_pathogenic, Uncertain_significance, Benign, Likely_benign
 
-For PRS analysis (optional):
+For PGx and PRS analysis (optional):
+    --run-pgx                       Run PGx analysis or not (optional), default is no
+                                    Options: yes, no
     --run-prs                       Run PRS analysis or not (optional), default is no
                                     Options: yes, no
     --run-imputation                Run genotype imputation or not (optional). This argument is only for --run-prs.
@@ -107,6 +109,9 @@ PGSID=""
 PGPID=""
 EFOID=""
 
+# PGx parameters
+RUNPGX="no"
+
 # Parse command line arguments
 while [[ $# -gt 0 ]]; do
     case $1 in
@@ -149,6 +154,10 @@ while [[ $# -gt 0 ]]; do
         --newborn)
             NEWBORN=true
             shift
+            ;;
+        --run-pgx)
+            RUNPGX="$2"
+            shift 2
             ;;
         --run-prs)
             RUNPRS="$2"
@@ -316,12 +325,13 @@ echo "Sample Gender: $GENDER"
 echo "VEP fork: $FORK"
 echo "Bcftools threads: $THREADS"
 
-echo "=== PRS Parameters ==="
+echo "=== PRS and PGx Parameters ==="
 echo "Run PRS analysis: $RUNPRS"
 echo "Run Imputation: $RUNIMPUTATION"
 echo "PGS IDs: $PGSID"
 echo "PGS Publications: $PGPID"
 echo "PGS EFO Traits: $EFOID"
+echo "Run PGx analysis: $RUNPGX"
 
 echo "=== Threshold Parameters ==="
 echo "Only ClinVar: $ONLY_CLINVAR"
@@ -413,38 +423,44 @@ conda run -n varxomics Rscript $SCRIPTS/RScripts/Newborn_Single20251203.R $sampl
 mv $OUTPUT_DIR/${INPUT_SAMPLE}.txt $OUTPUT_DIR/Results/
 
 ### Step 6: PGx by PharmCat
-pharmcat="/mnt/nas/Genomics/Genome/FamilyRisk/tools/pharmcat/pharmcat-3.1.1-all.jar"
-pharmcat_preprocessor="/mnt/nas/Genomics/Genome/FamilyRisk/tools/pharmcat/preprocessor/pharmcat_vcf_preprocessor"
-preprocessor_ref="/mnt/nas/Genomics/Genome/FamilyRisk/tools/pharmcat/reference.fna.bgz"
-preprocessor_position="/mnt/nas/Genomics/Genome/FamilyRisk/tools/pharmcat/pharmcat_positions_3.1.1.vcf.bgz"
+if [[ "$RUNPGX" == "no" ]]; then
+    echo "Skipping PGx analysis as --run-pgx is not set to 'yes'."
+    exit 0
+else
+    pharmcat="/mnt/nas/Genomics/Genome/FamilyRisk/tools/pharmcat/pharmcat-3.1.1-all.jar"
+    pharmcat_preprocessor="/mnt/nas/Genomics/Genome/FamilyRisk/tools/pharmcat/preprocessor/pharmcat_vcf_preprocessor"
+    preprocessor_ref="/mnt/nas/Genomics/Genome/FamilyRisk/tools/pharmcat/reference.fna.bgz"
+    preprocessor_position="/mnt/nas/Genomics/Genome/FamilyRisk/tools/pharmcat/pharmcat_positions_3.1.1.vcf.bgz"
 
-PHARMCAT_PREPROCESSED_VCF="${OUTPUT_DIR}/${INPUT_SAMPLE}_biallelic_nodup_pass.preprocessed.vcf.bgz"
-echo $PHARMCAT_PREPROCESSED_VCF
+    PHARMCAT_PREPROCESSED_VCF="${OUTPUT_DIR}/${INPUT_SAMPLE}_biallelic_nodup_pass.preprocessed.vcf.bgz"
+    echo $PHARMCAT_PREPROCESSED_VCF
 
-mkdir -p ${OUTPUT_DIR}/PGx
-mkdir -p ${OUTPUT_DIR}/Results/PGx_Reports
+    mkdir -p ${OUTPUT_DIR}/PGx
+    mkdir -p ${OUTPUT_DIR}/Results/PGx_Reports
 
-# normalized by pharmcat preprocessor
-conda run -n vep114 $pharmcat_preprocessor \
-    -vcf $OUTPUT_DIR/${INPUT_VCF_BIALLELIC_NODUP_PASS} \
-    -refFna $preprocessor_ref \
-    -refVcf $preprocessor_position
+    # normalized by pharmcat preprocessor
+    conda run -n vep114 $pharmcat_preprocessor \
+        -vcf $OUTPUT_DIR/${INPUT_VCF_BIALLELIC_NODUP_PASS} \
+        -refFna $preprocessor_ref \
+        -refVcf $preprocessor_position
 
-# pharmcat step A  
-java -jar $pharmcat \
-    -matcher -vcf "$PHARMCAT_PREPROCESSED_VCF" \
-    -phenotyper -o "${OUTPUT_DIR}/PGx/" \
-    -research cyp2d6 -v
+    # pharmcat step A  
+    java -jar $pharmcat \
+        -matcher -vcf "$PHARMCAT_PREPROCESSED_VCF" \
+        -phenotyper -o "${OUTPUT_DIR}/PGx/" \
+        -research cyp2d6 -v
 
-# pharmcat step B
-java -jar $pharmcat \
-    -reporter -ri ${OUTPUT_DIR}/PGx/*.phenotype.json \
-    -o "${OUTPUT_DIR}/Results/PGx_Reports" -reporterJson -reporterHtml -v
+    # pharmcat step B
+    java -jar $pharmcat \
+        -reporter -ri ${OUTPUT_DIR}/PGx/*.phenotype.json \
+        -o "${OUTPUT_DIR}/Results/PGx_Reports" -reporterJson -reporterHtml -v
 
-# Remove intermediate files
-rm -f "${PHARMCAT_PREPROCESSED_VCF}"*
-rm -f "${OUTPUT_DIR}/PGx"/*missing_pgx_var.vcf
-rm -f "${OUTPUT_DIR}"/*missing_pgx_var.vcf
+    # Remove intermediate files
+    rm -f "${PHARMCAT_PREPROCESSED_VCF}"*
+    rm -f "${OUTPUT_DIR}/PGx"/*missing_pgx_var.vcf
+    rm -f "${OUTPUT_DIR}"/*missing_pgx_var.vcf
+
+fi
 
 ### Step 7: Run PRS analysis if required
 if [[ "$RUNPRS" == "yes" ]]; then
